requirements.txt
APScheduler==3.9.1.post1
asgiref==3.6.0
autopep8==2.0.1
beautifulsoup4==4.12.3
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
commonmark==0.9.1
cryptography==42.0.8
diff-match-patch==20230430
Django==4.1.5
django-cors-headers==3.13.0
django-crontab==0.7.1
django-debug-toolbar==4.2.0
django-import-export==4.0.8
django-rest-knox==4.2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
frozendict==2.4.4
html5lib==1.1
idna==3.4
lxml==5.2.2
multitasking==0.0.11
mysqlclient==2.1.1
numpy==1.24.1
openpyxl==3.1.4
pandas==1.5.2
peewee==3.17.5
platformdirs==4.2.2
pycodestyle==2.10.0
pycparser==2.22
Pygments==2.14.0
pyotp==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
requests==2.32.3
rich==12.6.0
six==1.16.0
smartapi-python==1.3.5
soupsieve==2.5
sqlparse==0.4.3
tablib==3.5.0
tomli==2.0.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
webencodings==0.5.1
websocket-client==1.4.2
yfinance==0.2.40
